// src/modules/builder/shared/lib/models/BuilderMarketplaceCategory.ts
export enum BuilderMarketplaceCategory {
  AUTOMATISATION = 'automatisation',
  TELEPHONY = 'telephony',
  WAZZUP = 'wazzup',
  TILDA = 'tilda',
  WORDPRESS = 'wordpress',
  TWILIO = 'twilio',
  FB_MESSENGER = 'fb_messenger',
  SALESFORCE = 'salesforce',
  MAKE = 'make',
  APIX_DRIVE = 'apix_drive',
  ALBATO = 'albato',
  ONE_C = 'one_c',
}